+++ lib/src/helper/monitor_pop_route.dart
@@ -14,10 +14,9 @@ class MonitorPopRoute with WidgetsBindingObserver {
   static MonitorPopRoute get instance => _instance ??= MonitorPopRoute._();
 
   MonitorPopRoute._() {
-    if (WidgetsBinding.instance == null)
-      WidgetsFlutterBinding.ensureInitialized();
+    WidgetsFlutterBinding.ensureInitialized();
     //to prevent ensureInitialized changes, still use WidgetsBinding.instance
-    WidgetsBinding.instance!.addObserver(this);
+    WidgetsBinding.instance.addObserver(this);
   }
 
   @override