+++ lib/src/widget/toast_helper.dart
@@ -23,7 +23,7 @@ class _ToastHelperState extends State<ToastHelper> with WidgetsBindingObserver {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance?.addObserver(this);
+    WidgetsBinding.instance.addObserver(this);
     _dealKeyboard();
   }
 
@@ -45,7 +45,7 @@ class _ToastHelperState extends State<ToastHelper> with WidgetsBindingObserver {
 
   @override
   void dispose() {
-    WidgetsBinding.instance?.removeObserver(this);
+    WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }